+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This release adds a few new config options, for your convenience:
 
+- `scroll` controls how many lines to jump by when paging up/down.
+  If set to 0 (the new default), you'll jump by an entire screen.
 - `autoplay` controls whether you'll be prompted to play media files
   or not. By default it's false, but one might find it handy to set
   to `true` if hosting, say, a Gopher-powered music server.